modified redirect after submission, user gets warning on invalid subsmision, dynamic rendering of problem on live coding
