+++ client/src/containers/UserContainer.jsx
@@ -12,7 +12,7 @@ import { connect } from 'react-redux';
 
 import { fetchAllCompanyCalendars, fetchCandidateCalendar, fetchInitialChallenge, fetchCandidateInfo, updateCandidateSkills, updateCandidateGithub, saveCandidateCalendar, deleteCandidateSchedule, saveResults, currentCompanyCalendar, fetchCandidateInitialResults } from '../actions/userActions';
 import { fetchCompanySchedule, fetchCompanyResults } from '../actions/adminActions' ;
- 
+
 import axios from 'axios';
 
 class UserContainer extends Component {
